+++ ChangeLog.md
@@ -5,7 +5,6 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 # TO-DO List
 
 - Add command `/log copy <player>` or `/logCopy <player>`, to copy logger configurations from other players.
-- Add field `NAME` for all the logHelpers.
 - Display the actual option when setting logger with default options.
     - `Subscribed to <logger> \(<option>\)`
 - Bump to release versions after merging pathfinding-logger?
@@ -16,3 +15,7 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 
 - Fix a display bug about symbol `%` of command `/spawn` on the server console.
 - Add `pathFinding` logger, displaying entity AI paths with particle lines.
+
+# Code Details
+
+- Add field `NAME` for logHelpers of `lightCheck` and `rngManip`.